requirements.txt
# This is an implicit value, here for clarity
--index-url https://pypi.python.org/simple/

pyzmq==14.0
kazoo==2.0
